:truck: Move set, unset and update into object namespace, remove core namespace
