+++ src/array/push.js
@@ -1,5 +1,5 @@
 import toArray from 'lodash/toArray'
-import update from '../core/update'
+import update from '../object/update'
 
 /**
  * Replaces by a new array containing the elements of the former array and one or more elements added to the end.
@@ -15,7 +15,7 @@ import update from '../core/update'
  * push({ nested: { prop: [1, 2] } }, 'nested.prop', 3, 4) // => { nested: { prop: [1, 2, 3, 4] } }
  * @see {@link https://lodash.com/docs#toArray|lodash.toArray} for more information.
  * @see {@link https://mdn.io/Array.prototype.push|Array.prototype.push} for more information.
- * @see {@link core.update|update} for more information.
+ * @see {@link object.update|update} for more information.
  * @since 0.1.7
  */
 export default update((value, ...values) => {
