--- src/core/index.js
@@ -1,14 +1,0 @@
-import set from './set'
-import unset from './unset'
-import update from './update'
-
-/**
- * Core functions.
- * @namespace core
- * @since 0.1.5
- */
-export {
-  set,
-  unset,
-  update,
-}
