+++ src/index.js
@@ -1,6 +1,6 @@
 export * from './array'
 export * from './collection'
-export * from './core'
+export * from './object'
 export * from './lang'
 export * from './math'
 export * from './object'
