+++ src/math/add.js
@@ -1,5 +1,5 @@
 import add from 'lodash/add'
-import update from '../core/update'
+import update from '../object/update'
 
 /**
  * Replaces by the addition of the former number and the given number.
@@ -11,7 +11,7 @@ import update from '../core/update'
  * @return {Object} Returns the updated object.
  * @example add({ nested: { prop: 2 } }, 'nested.prop', 4) // => { nested: { prop: 6 } }
  * @see {@link https://lodash.com/docs#add|lodash.add} for more information.
- * @see {@link core.update|update} for more information.
+ * @see {@link object.update|update} for more information.
  * @since 0.1.7
  */
 export default update(add)
