+++ src/object/index.js
@@ -1,6 +1,9 @@
 import assign from './assign'
 import mapValues from './mapValues'
 import pickBy from './pickBy'
+import set from './set'
+import unset from './unset'
+import update from './update'
 
 /**
  * Array functions.
@@ -11,4 +14,7 @@ export {
   assign,
   mapValues,
   pickBy,
+  set,
+  unset,
+  update,
 }
