+++ src/object/pickBy.js
@@ -1,6 +1,6 @@
 import { lodashFpConvertOptions } from '../consts'
 import pickBy from 'lodash/fp/pickBy'
-import update from '../core/update'
+import update from '../object/update'
 
 const rawPickBy = pickBy.convert(lodashFpConvertOptions)
 
@@ -14,7 +14,7 @@ const rawPickBy = pickBy.convert(lodashFpConvertOptions)
  * @param {function} [predicate={@link https://lodash.com/docs#identity|lodash.identity}] The function invoked per iteration.
  * @return {Object} Returns the updated object.
  * @example pickBy({ nested: { a: 1, b: 2, c: 3, d: 4 } }, 'nested', v => v < 3) // => { nested: { a: 1, b: 2 } }
- * @see {@link core.update|update} for more information.
+ * @see {@link object.update|update} for more information.
  * @see {@link https://lodash.com/docs#pickBy|lodash.pickBy} for more information.
  * @see {@link https://lodash.com/docs#identity|lodash.identity} for more information.
  * @since 0.1.12
