+++ src/object/set.js
@@ -5,7 +5,7 @@ import { lodashFpConvertOptions } from '../consts'
  * Sets the value at <code>path</code> of <code>object</code>.
  * This is the <code>lodash/fp</code> <code>set</code>, with no arguments rearranging and no currying.
  * @function set
- * @memberof core
+ * @memberof object
  * @param {Object} object The object to modify.
  * @param {Array|string} path The path of the property to set.
  * @param {*} value The value to set.
