+++ src/object/set.spec.js
