+++ src/object/update.spec.js
